requirements.txt
# AI/LLM
openai>=1.0.0

# Environment variables
python-dotenv>=0.19.0

# Web3 interaction
web3>=6.0.0

# Data handling
pydantic>=2.0.0
python-decimal>=1.0.0

# Async support
aiohttp>=3.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0

# Development
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0